Add in a new Book Controller.
This is necessary becuase I've chosen to encode the books data into the content field, to avoid extraneous calls to the API, and I need to have the raw values returned. I could do an extra call to the API to return a call for meta data, but the api for meta data is hard to use, since there isn't a proper PUT call which can edit OR add, as PUT should be able to do. Because of this, there are several calls to the meta api needed, to first verify the meta value is there, if so PUT to it, otherwise POST to it. This is just overly complex and since I need this plugin anyway, might as well make this small modification to make the rest of the code more straight-forward.